Restructure assemblies
