+++ Assets/GraphicTests/com.alelievr.mixture.tests.asmdef
@@ -2,13 +2,13 @@
     "name": "com.alelievr.mixture.tests",
     "rootNamespace": "",
     "references": [
-        "UnityEngine.TestRunner",
-        "Mixture.Runtime",
-        "com.alelievr.NodeGraphProcessor.Runtime",
-        "UnityEngine.TestTools.Graphics",
-        "UnityEditor.TestRunner",
-        "Unity.ShaderGraph.GraphicsTests",
-        "UnityEditor.TestTools.Graphics"
+        "GUID:27619889b8ba8c24980f49ee34dbb44a",
+        "GUID:fe40ba3f519cd4795b386f0ccc63a5a7",
+        "GUID:ca937d03ee5dd4d699091438dc0f3ae6",
+        "GUID:c081bc530f560634bb5c21d4b323a7f1",
+        "GUID:0acc523941302664db1f4e527237feb3",
+        "GUID:31660b124da7540378315e14658e3cc6",
+        "GUID:e18141520846dcc44b725b2f74e91229"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
