+++ Packages/com.alelievr.mixture/Editor/CustomTextureShaderGraph/CustomTextureNodes.cs
@@ -1,4 +1,3 @@
-#if MIXTURE_SHADERGRAPH
 using UnityEngine;
 using UnityEditor.ShaderGraph;
 using UnityEditor.Graphing;
@@ -244,4 +243,3 @@ namespace Mixture
             => NeededCoordinateSpace.World;
     }
 }
-#endif
