+++ Packages/com.alelievr.mixture/Editor/CustomTextureShaderGraph/FullScreenPassTarget.cs
@@ -1,4 +1,3 @@
-#if MIXTURE_SHADERGRAPH
 using System;
 using System.Linq;
 using System.Collections.Generic;
@@ -219,7 +218,7 @@ namespace Mixture
         public override bool WorksWithSRP(RenderPipelineAsset scriptableRenderPipeline) => true; 
     }
 
-    static class FullscreePasses
+    static class FullscreenPasses
     {
         public static PassDescriptor CustomRenderTexture = new PassDescriptor
         {
@@ -309,4 +308,3 @@ namespace Mixture
 
     }
 }
-#endif
