+++ Packages/com.alelievr.mixture/Editor/CustomTextureShaderGraph/LegacyMasterNodes.cs
@@ -1,9 +1,4 @@
-#if MIXTURE_SHADERGRAPH
 using System;
-using System.Linq;
-using UnityEditor.Graphing;
-using UnityEditor.ShaderGraph.Drawing.Controls;
-using UnityEngine;
 using UnityEditor.ShaderGraph;
 using UnityEditor.ShaderGraph.Legacy;
 
@@ -25,4 +20,3 @@ namespace Mixture
         public const int ColorSlotId = 0;
     }
 }
-#endif
