+++ Packages/com.alelievr.mixture/Editor/CustomTextureShaderGraph/MixtureShaderGraphCallbacks.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 41adabb070894ec4bbebb64a5eceece0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
