+++ Packages/com.alelievr.mixture/Editor/CustomTextureShaderGraph/com.alelievr.Mixture.Editor.CustomTextureShaderGraph.asmref
@@ -1,0 +1,3 @@
+{
+    "reference": "GUID:be0903cd8e1546f498710afdc59db5eb"
+}
