+++ Packages/com.alelievr.mixture/Editor/CustomTextureShaderGraph/com.alelievr.Mixture.Editor.CustomTextureShaderGraph.asmref.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 75da831077dc0e14a97b95061e6f4bdc
+AssemblyDefinitionReferenceImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
