+++ Packages/com.alelievr.mixture/Editor/com.alelievr.Mixture.Editor.asmdef
@@ -1,14 +1,14 @@
 {
-    "name": "Unity.ShaderGraph.GraphicsTests",
+    "name": "com.alelievr.Mixture.Editor",
     "rootNamespace": "",
     "references": [
-        "Unity.ShaderGraph.Editor",
-        "Mixture.Runtime",
-        "com.alelievr.NodeGraphProcessor.Runtime",
-        "com.alelievr.NodeGraphProcessor.Editor",
-        "Unity.RenderPipelines.HighDefinition.Runtime",
-        "Unity.RenderPipelines.Core.Editor",
-        "Unity.RenderPipelines.Core.Runtime"
+        "GUID:fe40ba3f519cd4795b386f0ccc63a5a7",
+        "GUID:ca937d03ee5dd4d699091438dc0f3ae6",
+        "GUID:a432eb0a70c5c4a37ab50d59671586f8",
+        "GUID:457756d89b35d2941b3e7b37b4ece6f1",
+        "GUID:3eae0364be2026648bf74846acb8a731",
+        "GUID:df380645f10b7bc4b97d4f5eb6303d95",
+        "GUID:be0903cd8e1546f498710afdc59db5eb"
     ],
     "includePlatforms": [
         "Editor"
