+++ Packages/com.alelievr.mixture/Runtime/com.alelievr.Mixture.Runtime.asmdef
@@ -1,5 +1,5 @@
 {
-    "name": "Mixture.Runtime",
+    "name": "com.alelievr.Mixture.Runtime",
     "rootNamespace": "",
     "references": [
         "GUID:ca937d03ee5dd4d699091438dc0f3ae6",
