+++ Packages/manifest.json
@@ -10,7 +10,7 @@
   ],
   "dependencies": {
     "com.alelievr.mixture": "file:com.alelievr.mixture",
-    "com.alelievr.node-graph-processor": "1.3.0",
+    "com.alelievr.node-graph-processor": "https://github.com/alelievr/NodeGraphProcessor.git?path=/Assets/com.alelievr.NodeGraphProcessor",
     "com.unity.2d.sprite": "1.0.0",
     "com.unity.2d.tilemap": "1.0.0",
     "com.unity.barracuda": "1.0.4",
