+++ Packages/packages-lock.json
@@ -11,10 +11,11 @@
       }
     },
     "com.alelievr.node-graph-processor": {
-      "version": "file:C:/Users/Antoine Lelievre/Github/NodeGraphProcessor/Assets/com.alelievr.NodeGraphProcessor",
+      "version": "https://github.com/alelievr/NodeGraphProcessor.git?path=/Assets/com.alelievr.NodeGraphProcessor",
       "depth": 0,
-      "source": "local",
-      "dependencies": {}
+      "source": "git",
+      "dependencies": {},
+      "hash": "bc71d48ba9dfc7e7062779d6d12b9ef269e521b1"
     },
     "com.autodesk.fbx": {
       "version": "4.0.0-pre.1",
